virt-controller, template: move the wrapExecProbeWithVirtProbe func
